+++ examples/updates/update_iam_service_account/providers.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 5.0"
+      version = "~> 5.0"
     }
   }
 }
@@ -16,6 +16,6 @@ provider "google" {
     send_after = "0s"
   }
 
-  iam_custom_endpoint                    = "http://localhost:8080/v1/"
-  cloud_resource_manager_custom_endpoint = "http://localhost:8080/v1/"
+  iam_custom_endpoint                    = "http://localhost:8080/"
+  cloud_resource_manager_custom_endpoint = "http://localhost:8080/"
 }