+++ examples/working/gke_cluster/main.tf
@@ -24,6 +24,8 @@ resource "google_container_cluster" "main" {
   # but we want to manage them separately.
   remove_default_node_pool = true
   initial_node_count       = 1
+
+  deletion_protection = false
 }
 
 resource "google_container_node_pool" "primary" {